"refactor(US01): replace double-quotes with single-quotes"
