+++ src/app/utils/constants/route.consts.ts
@@ -1,16 +1,16 @@
-// import {RouteConfig, RouteConfigs} from "../interfaces/route.interfaces";
-//
-// const home: RouteConfig = {
-//   name: "home",
-//   path: ""
-// }
-//
-// const postPage: RouteConfig = {
-//   name: "post-page",
-//   path: "post/:id"
-// }
-//
-// export const ROUTE_CONFIGS: RouteConfigs = {
-//   home,
-//   postPage,
-// }
+import {RouteConfig, RouteConfigs} from '../interfaces/route.interfaces';
+
+const home: RouteConfig = {
+  name: 'home',
+  path: ''
+}
+
+const postPage: RouteConfig = {
+  name: 'post-page',
+  path: 'post/:id'
+}
+
+export const ROUTE_CONFIGS: RouteConfigs = {
+  home,
+  postPage,
+}
