+++ src/app/utils/interfaces/route.interfaces.ts
@@ -1,10 +1,10 @@
-// export interface RouteConfig {
-//   name: string,
-//   path: string,
-//   fullpath?: string,
-//   param?: string,
-// }
-//
-// export interface RouteConfigs {
-//   [key: string]: RouteConfig;
-// }
+export interface RouteConfig {
+  name: string,
+  path: string,
+  fullpath?: string,
+  param?: string,
+}
+
+export interface RouteConfigs {
+  [key: string]: RouteConfig;
+}
